Extract Instruction etc. into instructions.py
